+++ src/restful/routes/authRouter.js
@@ -4,18 +4,6 @@ const { protect } = require("../../middleware");
 
 const router = Router();
 
-const rateLimit = require('express-rate-limit');
-
-// Define rate limiting options
-const limiter = rateLimit({
-  windowMs: 60 * 1000, // 1 minute
-  max: 5, // Max 5 requests per minute
-  message: "Too many requests from this IP, please try again later."
-});
-
-// Apply rate limiting middleware to all routes in the applications router
-router.use(limiter);
-
 router.get("/user", protect, AuthController.getUser);
 router.put("/user", protect, AuthController.updateUser);
