+++ src/restful/routes/opportunitiesRouter.js
@@ -5,18 +5,6 @@ const {
 
 const router = Router();
 
-const rateLimit = require('express-rate-limit');
-
-// Define rate limiting options
-const limiter = rateLimit({
-  windowMs: 60 * 1000, // 1 minute
-  max: 5, // Max 5 requests per minute
-  message: "Too many requests from this IP, please try again later."
-});
-
-// Apply rate limiting middleware to all routes in the applications router
-router.use(limiter);
-
 router.get("/", OpportunitiesController.getAllOpportunities);
 router.get("/id/:user_id", OpportunitiesController.getAllOpportunitiesByUserId);
 router.get("/status/:status", OpportunitiesController.getOpportunitiesByStatus);
